refactor: 기존 HttpRequest 클래스 분리
- HttpRequest가 갖고 있던, 요청에 대한 처리를 HttpRequestProcessor
  클래스로 분리했습니다.
- HttpRequest에는 요청 메시지의 정보인 start line, headers, body 갖고
  있도록 하고, 관련 로직들을 수행하게 할 예정입니다.
